FuncionQR Lista

+++ MVVM/Models/Item.cs
@@ -6,7 +6,7 @@
         public string name { get; set; }
         public string description { get; set; }
         public string linkImg { get; set; }
-        public float price { get; set; }
+        public int price { get; set; }
 
     }
 }
